Extract section display components into separate file so we can use in both server and client components
